Added update energy, updated layout to fit better on different phone sizes

+++ STEPN Sidekick/ActivitySettings.swift
@@ -37,7 +37,7 @@ struct ActivitySettings: View {
     @State private var voiceAlertsSpeedType: Int = speedAlertsBoth
     @State private var voiceAlertsTime: Bool = true
     @State private var voiceAlertsMinThirty: Bool = true
-    @State private var energy: Float = 2.0
+    @State private var energy: Double = 2.0
     @State private var shoeTypeIterator: Int = walker
     @State private var customMinSpeed: Float = 0
     @State private var customMaxSpeed: Float = 0
@@ -399,7 +399,7 @@ struct ActivitySettings: View {
                                             
                                             Spacer()
                                             
-                                            Text(getMinString(energy: Float(energyString) ?? 0.0))
+                                            Text(getMinString(energy: Double(energyString) ?? 0.0))
                                                 .font(Font.custom(fontButtons, size: 16))
                                                 .foregroundColor(Color("Gandalf"))
                                                 .frame(minWidth: 150, maxWidth: 160)
@@ -523,7 +523,7 @@ struct ActivitySettings: View {
                                                 if Double(energyString) ?? 0 == 0 {
                                                     noEnergyAlert = true
                                                 } else {
-                                                    energy = Float(energyString) ?? 0.2
+                                                    energy = Double(energyString) ?? 0.2
                                                     self.startSpeedTracker = true
                                                 }
                                                 
@@ -540,7 +540,7 @@ struct ActivitySettings: View {
                                                     if Double(energyString) ?? 0 == 0 {
                                                         noEnergyAlert = true
                                                     } else {
-                                                        energy = Float(energyString) ?? 0.2
+                                                        energy = Double(energyString) ?? 0.2
                                                         self.startSpeedTracker = true
                                                     }
                                                 }))
@@ -579,7 +579,7 @@ struct ActivitySettings: View {
         return buttonText
     }
     
-    func getMinString(energy: Float) -> String {
+    func getMinString(energy: Double) -> String {
         var energyInMins: Int = Int(energy * 5)
         var minString: String
